Make Navigation bar a component and add it to pages
