+++ lib/pages/home.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:hiking4nerds/components/hikingmapbox.dart';
-import 'package:hiking4nerds/styles.dart';
+import 'package:hiking4nerds/components/navbar.dart';
 
 class Home extends StatefulWidget {
   @override
@@ -8,15 +8,6 @@ class Home extends StatefulWidget {
 }
 
 class _HomeState extends State<Home> {
-
-    static const TextStyle optionStyle = TextStyle(fontSize: 30, fontWeight: FontWeight.bold);
-
-    void _onItemTapped(int index) {
-      List<String> options = ['/help', 'routeinfo', '/settings'];
-
-      if(index < options.length || options[index] != 'routeinfo') Navigator.pushNamed(context, options[index]);
-      // else RouteInfo();
-    }
 
   @override
   Widget build(BuildContext context) {
@@ -30,37 +21,7 @@ class _HomeState extends State<Home> {
         children: <Widget>[
           MapWidget(),      
         ]),
-        floatingActionButtonLocation: FloatingActionButtonLocation.centerDocked,
-      floatingActionButton: FloatingActionButton(
-        onPressed: () { Navigator.pushNamed(context, '/routesetup');},
-        tooltip: 'Route Setup',
-        child: Icon(Icons.find_replace),
-        backgroundColor: htwGreen,
-        elevation: 2.0,
-      ),
-      bottomNavigationBar: BottomNavigationBar(
-        items: const <BottomNavigationBarItem> [
-          BottomNavigationBarItem(
-            icon: Icon(Icons.help_outline),
-            title: Text('Help'),
-          ),
-          BottomNavigationBarItem(
-            icon: Icon(Icons.find_replace),
-            title: Text('Route'),
-          ),
-          BottomNavigationBarItem(
-            icon: Icon(Icons.settings),
-            title: Text('Settings'),
-          ),
-          BottomNavigationBarItem(
-            icon: Icon(Icons.save),
-            title: Text('Last Routes'),
-          ),
-        ],
-        onTap: _onItemTapped,
-        unselectedItemColor: Colors.grey,
-        selectedItemColor: Colors.grey,
-      ),
+      bottomNavigationBar: Navbar(),
     );
   }
 }
